pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
backend-path = ["."]

[project]
name = "ahc-agent-cli"
version = "0.1.0"
authors = [
    { name = "AHCAgent Team", email = "ahcagent@example.com" },
]
description = "A CLI tool for solving AtCoder Heuristic Contest problems using AlphaEvolve"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT License" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "litellm>=0.1.0",
    "docker>=6.0.0",
    "aiohttp>=3.8.0",
    "tqdm>=4.64.0",
    "colorama>=0.4.4",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.4.4",
  "pre-commit<3.7",
  "pytest>=8.2.0",
  "pytest-asyncio>=0.20.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/ahc-agent-cli"

[project.scripts]
ahc-agent = "ahc_agent_cli.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
"ahc_agent_cli" = ["templates/*.cpp", "templates/*.md", "config/*.yaml"]

[tool.ruff]
# Pythonバージョンを指定
# target-version = "py38"
line-length = 128
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "S", "BLE", "B", "A", "C4", "T10", "ISC", "ICN", "PIE", "PYI", "PT", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "NPY", "PERF", "RUF"]
ignore = ["ISC001", "B023", "PERF203", "BLE001", "S108", "PT011", "S603", "S607"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["ahc_agent_cli"]
force-sort-within-sections = true

[tool.ruff.lint.per-file-ignores]
  "tests/**/*.py" = ["S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
